backend/requirements.txt
Django==4.2

django-cors-headers==4.4.0

django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.4
ninja-schema==0.13.6

djangorestframework==3.15.2
coreapi==2.3.3
djangorestframework-simplejwt==5.3.1